normaliz/requirements.txt
Flask
flask-restful
PyYAML
requests